pyproject.toml
[tool.poetry]
name = "djangosocialnetworkbot"
version = "0.1.0"
description = ""
authors = ["mmeerrccyy <taras.mercury@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
